Refactored solution for day One
