+++ AdventOfCode2021/DayOne/PartOneSolution.cs
@@ -1,0 +1,20 @@
+﻿namespace DayOne;
+
+internal class PartOneSolution
+{
+    public int Run(IList<int> depthMeasurements)
+    {
+        var previousMeasurement = depthMeasurements.First();
+        int biggerMeasurements = 0;
+
+        foreach (var measurement in depthMeasurements)
+        {
+            if (measurement > previousMeasurement)
+            {
+                biggerMeasurements++;
+            }
+            previousMeasurement = measurement;
+        }
+        return biggerMeasurements;
+    }
+}
